Fix error thrown when validating certain color strings
parseCssColor checks `namedColors[input]` to see if the given color
string is a predetermined named color. If "__proto__" or any other key
found in Object.prototype was used, the evaluated value was truthy, and
the function attempted to destructure the returned value, which throws
an error.

This commit fixes that issue by using getOwn() to access only own
properties of the namedColors object.