Fix color coercion expressions returning array of nulls
Color coercion (Coercion class) uses EvaluationContext.parseColor to
parse color strings. EvaluationContext has an internal cache for color
strings, which it checks before parsing to skip parsing the same value
multiple times.

If a color coercion expression was passed "__proto__" or some other key
present in Object.prototype, the parseColor function would skip parsing
altogether as this._parseColorCache["__proto__"] evaluated to a truthy
value, and thus parseColor thought the value was cached and returned the
evaluated value directly.

This caused the entire coercion expression to evaluate to the value of
Object.prototype, which in the end resulted in the color being
[null, null, null, null] somewhere in maplibre-gl-js.

Fixed by creating the cache with a null prototype, instead of an empty
object literal.