Fix prototype keys causing crash in color parsing and invalid colors in color coercion expressions (#1036)
* Fix error thrown when validating certain color strings

parseCssColor checks `namedColors[input]` to see if the given color
string is a predetermined named color. If "__proto__" or any other key
found in Object.prototype was used, the evaluated value was truthy, and
the function attempted to destructure the returned value, which throws
an error.

This commit fixes that issue by using getOwn() to access only own
properties of the namedColors object.

* Fix color coercion expressions returning array of nulls

Color coercion (Coercion class) uses EvaluationContext.parseColor to
parse color strings. EvaluationContext has an internal cache for color
strings, which it checks before parsing to skip parsing the same value
multiple times.

If a color coercion expression was passed "__proto__" or some other key
present in Object.prototype, the parseColor function would skip parsing
altogether as this._parseColorCache["__proto__"] evaluated to a truthy
value, and thus parseColor thought the value was cached and returned the
evaluated value directly.

This caused the entire coercion expression to evaluate to the value of
Object.prototype, which in the end resulted in the color being
[null, null, null, null] somewhere in maplibre-gl-js.

Fixed by creating the cache with a null prototype, instead of an empty
object literal.

* Update CHANGELOG.md

* Group prototype key related fixes in CHANGELOG.md

* Use Map instead of an object in EvaluationContext's cache